requirements.txt
Django==1.9.8
django-embed-video==1.1.0
django-contact-form==1.2
django-parler==1.6.5